+++ flaskr/schema.sql
@@ -34,3 +34,7 @@ create table notification (
         On Delete Cascade
         On Update Cascade
 );
+
+INSERT INTO user (email, adminStatus, password) VALUES ('test@hotmail.com', 1, 'test');
+INSERT INTO notification (userid, minMagnitude, maxMagnitude, latitude, longitude, location, radius, attributes) VALUES (1, 0, 10, 0, 0, 'test', 0, 'test');
+INSERT INTO earthquake (title, eventTime, magnitude, latitude, longitude, depth, url) VALUES ('test', '2020-01-01', 0, 0, 0, 0, 'test');
