+++ flaskr/templates/manageNotifications.html
@@ -35,11 +35,11 @@
                     <span>OR</span>
                     <div class="col-2">
                         <label for="latitude" class="form-label">Latitude</label>
-                        <input type="number" class="form-control" id="latitude">
+                        <input type="number" class="form-control" id="latitude" step="0.1">
                     </div>
                     <div class="col-2">
                         <label for="longitude" class="form-label">Longitude</label>
-                        <input type="number" class="form-control" id="longitude">
+                        <input type="number" class="form-control" id="longitude" step="0.1">
                     </div>
                 </div>
                 <div class="row justify-content-center mt-5">
@@ -47,15 +47,15 @@
                         <label for="radius" class="form-label">Radius</label>
                         <select type="number" class="form-select" id="radius" name="radius">
                             <option value="1">1km</option>
-                            <option value="1">2km</option>
-                            <option value="1">5km</option>
-                            <option value="1">10km</option>
-                            <option value="1">20km</option>
-                            <option value="1">40km</option>
-                            <option value="1">60km</option>
-                            <option value="1">80km</option>
-                            <option value="1">100km</option>
-                            <option value="1">250km</option>
+                            <option value="2">2km</option>
+                            <option value="5">5km</option>
+                            <option value="10">10km</option>
+                            <option value="20">20km</option>
+                            <option value="40">40km</option>
+                            <option value="60">60km</option>
+                            <option value="80">80km</option>
+                            <option value="100">100km</option>
+                            <option value="250">250km</option>
                         </select>
                     </div>
                 </div>
